CSCI 271 Lab 4
This is just a rehash of lab 3 with a few do-while loops.  A couple of the do-while loops has the program check if the user entered valid test scores or gpas, and reprompts them for this information if not.  The teacher had us redo the previous lab because we all used while loops instead of do-while loops for the menu.