CSCI 271 Lab 5
This lab involves a header file with two functions, one that converts centimeters to inches, and another that converts inches to centimeters.  The main cpp file has code for a main menu where the user can select what conversion they want the program to do.  The cpp file only has code for the main menu, and each choice calls the appropriate function in the header file.  There is also a counter that counts how many times the user accesses the main menu, which each function increases when called.  After the user chooses to exit the program, the program tells the user how many times they accessed the main menu.