CSCI 271 Lab 6
This is the last lab for my CSCI 271 class.  This simply takes 5 student test scores from an imaginary class, stores them into an array, calculates the average, sorts the scores in ascending order and then presents all the information to the user.  The point of the lab was to practice using arrays and for loops for entering loads of information in arrays.