CSCI 272 Lab 1
This program gives the user options in a menu to have the program find the circumference and area of a circle; or the area and perimeter of a rectangle.  The main file utilizes a header file with classes allowing for the creation of each shape so that the area, perimeter or circumference may be found.  The accompanying implementation file defines each function that each class has in the header file and allows the main program to implement the functions.