CSCI 272 Lab 2
This lab is the same as the first one, except some of the operators in the main program are overloaded so that two areas of the same shape may be added together and output, or two areas of different shapes may be compared.  The source code for the first lab was used as a basis for all three files.