CSCI 272 Lab 3
This program is simply a program that takes in numbers as input to make a list, and then gives the user the option of sorting the list before printing it out.  The program uses try-catch blocks to check for invalid input, and issues error messages according to the type of exception occurring.  If the user enters a letter instead of a number, the program prints an error message and asks the user for a number again.  If the list is not made first before being sorted, the program prints out a message that the list is empty.  All possible exceptions will produce the same error messages.