CS 272 Lab 4
This lab implements a linked-list to hold data.  The program presents a menu for the user to do different things with the list.   The user may add an element, delete an element, search for an element in the list, and print out the list.  The list's nodes are all data structures with a pointer and an integer variable to hold data.  Everything that works in the program is all in one file, and no classes are used.